acfpasswd: new tool to set passwords from comman line
This tool allows users to set/reset an ACF password from command line.

It also allows user to syncronize the ACF password with the system
password so the ACF password becomes same as shell login password. This
requires that the username exists in both /etc/acf/passwd and /etc/shadow